draw_filters moved to main controller because it is app specific
